# storage: reduce allocations in MVCCIncrementalIterator