This changes some uses of Key() to UnsafeKey() to avoid unnecessary
allocations.  I've added a small benchmark that shows the decreased
number of allocations when iterating.

```
name                                            old time/op    new time/op    delta
MVCCIncrementalIterator/useTBI=true/ts=5-16       12.5ms ± 7%     9.0ms ± 7%  -28.02%  (p=0.000 n=57+26)
MVCCIncrementalIterator/useTBI=true/ts=480-16     2.36ms ± 8%    2.27ms ±13%   -4.12%  (p=0.000 n=52+25)
MVCCIncrementalIterator/useTBI=false/ts=5-16      6.32ms ± 7%    6.44ms ±11%     ~     (p=0.083 n=55+24)
MVCCIncrementalIterator/useTBI=false/ts=480-16    1.14ms ±10%    1.15ms ± 6%     ~     (p=0.390 n=54+25)

name                                            old alloc/op   new alloc/op   delta
MVCCIncrementalIterator/useTBI=true/ts=5-16        831kB ± 0%      32kB ± 0%  -96.20%  (p=0.000 n=57+23)
MVCCIncrementalIterator/useTBI=true/ts=480-16     60.5kB ± 0%    35.7kB ± 0%  -41.00%  (p=0.000 n=56+26)
MVCCIncrementalIterator/useTBI=false/ts=5-16      17.7kB ± 0%    17.7kB ± 0%     ~     (p=0.127 n=50+25)
MVCCIncrementalIterator/useTBI=false/ts=480-16    21.8kB ± 0%    21.8kB ± 0%     ~     (p=0.469 n=57+26)

name                                            old allocs/op  new allocs/op  delta
MVCCIncrementalIterator/useTBI=true/ts=5-16         103k ± 0%        0k ± 0%  -99.71%  (p=0.000 n=57+26)
MVCCIncrementalIterator/useTBI=true/ts=480-16      3.76k ± 0%     0.55k ± 0%  -85.28%  (p=0.000 n=57+26)
MVCCIncrementalIterator/useTBI=false/ts=5-16        22.0 ± 0%      22.0 ± 0%     ~     (all equal)
MVCCIncrementalIterator/useTBI=false/ts=480-16       277 ± 0%       277 ± 0%     ~     (all equal)
```

The benchmark uses data that I believe fits entirely into the cache,
which may explain why the TBI optimization is slower even when using a
timestamp that should exclude a large portion of the data.

Release note: None